WIP: Add Coverage
This patch adds a (almost, not entirely) working coverage system to Saddle, based on @0x/sol-coverage. This coverage system tracks all transactions sent to ganache and then after the tests are completed, it runs a trace on each of those transactions and then looks at the PC from those transactions and uses the source maps to figure out which lines were called. This is a *much better* way to run tests as it does not require changing every single file under test and changing a huge variety of behaviours. Currently, we're able to generate some basic coverage, we just need to expand which files it's tracking and make sure we can "combine" the coverage from each of the Jest workers.