Fix TypeScript Issues for Web3
This patch fixes typescript issues for the newest types of Web3 from the 1.2.4 branch. This should set us up to easily move to the 2.0.0 branch once that moves forward, since the types are more in line with these fixes. Additionally, we clean up the concept of a "Log" to move into a class so we can add functions like "show" and hopefully other useful features for users of the saddle trace library.