Rework Verify Command
This patch reworks the Verify command to accept an address instead of a contract name. This just seems easier overall than trying to manage the deployed mappings to use this function. We should sit and take a hard look at the command signatures and make sure they align with what we're looking for.